+++ filecheck.js
@@ -1,0 +1,7 @@
+// accepts a list of files and a callback
+function filecheck(list, cb) {
+
+  cb();
+}
+
+module.exports = filecheck;