+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## changelog for parallel-file-processing
 
 ### 1.0.0 (2016-05-11)
+
+#### changes:
+  - filecheck module created.